+++ azure-typescript/2_stack-advanced-topics/solutions/exercise-1.index.ts
@@ -5,9 +5,15 @@ import * as storage from "@pulumi/azure-native/storage";
 import * as web from "@pulumi/azure-native/web";
 import * as pulumi from "@pulumi/pulumi";
 
+
+// Get the password to use for SQL from config.
 const config = new pulumi.Config();
 const baseName = config.get("baseName") || pulumi.getStack();
 const username = config.get("sqlUsername") || "pulumi";
+// Exercise 1 //
+const pwd = config.requireSecret("sqlPassword");
+export const dbPassword = pwd;
+// Exercise 1 //
 
 const resourceGroup = new resource.ResourceGroup(`${baseName}-rg`);
 
@@ -102,7 +108,6 @@ const app = new web.WebApp(`${baseName}-webapp`, {
 });
 
 export const dbUserName = username;
-export const dbPassword = pwd;
 export const storageAccountKey = primaryStorageKey;
 export const appHostName = app.defaultHostName;
