+++ azure-typescript/2_stack-advanced-topics/solutions/exercise-3.index.ts
@@ -5,9 +5,15 @@ import * as storage from "@pulumi/azure-native/storage";
 import * as web from "@pulumi/azure-native/web";
 import * as pulumi from "@pulumi/pulumi";
 
+
+// Get the password to use for SQL from config.
 const config = new pulumi.Config();
 const baseName = config.get("baseName") || pulumi.getStack();
 const username = config.get("sqlUsername") || "pulumi";
+// Exercise 1 //
+const pwd = config.requireSecret("sqlPassword");
+export const dbPassword = pwd;
+// Exercise //
 
 const resourceGroup = new resource.ResourceGroup(`${baseName}-rg`);
 
@@ -22,7 +28,10 @@ const storageAccount = new storage.StorageAccount(`${baseName.toLowerCase()}sa`,
 
 const storageAccountKeys = pulumi.all([resourceGroup.name, storageAccount.name]).apply(([resourceGroupName, accountName]) =>
     storage.listStorageAccountKeys({ resourceGroupName, accountName }));
-const primaryStorageKey = storageAccountKeys.keys[0].value;
+
+// Exercise 2 //
+const primaryStorageKey = pulumi.secret(storageAccountKeys.keys[0].value);
+// Exercise 2 //
 
 const appServicePlan = new web.AppServicePlan(`${baseName}-asp`, {
     resourceGroupName: resourceGroup.name,
@@ -102,9 +111,16 @@ const app = new web.WebApp(`${baseName}-webapp`, {
 });
 
 export const dbUserName = username;
-export const dbPassword = pwd;
 export const storageAccountKey = primaryStorageKey;
 export const appHostName = app.defaultHostName;
+
+// Exercise 3 //
+// The wrong way - so one can see what the output looks like
+export const appUrlWrong = `https://${app.defaultHostName}`;
+// Correct way using "interpolate"
+export const appUrlUsingInterpolate = pulumi.interpolate`https://${app.defaultHostName}`;
+// Correct way using "apply"
+export const appUrlUsingApply = app.defaultHostName.apply(hostname => `https://${hostname}`);
 
 
 /////// Helper Function ///////
