+++ azure-typescript/2_stack-advanced-topics/solutions/exercise-4/config.ts
@@ -1,0 +1,6 @@
+import * as pulumi from "@pulumi/pulumi";
+
+export const config = new pulumi.Config();
+export const baseName = config.get("baseName") || pulumi.getStack();
+export const username = config.get("sqlUsername") || "pulumi";
+export const pwd = config.requireSecret("sqlPassword")
